Use return() instead of exit()
Previously, the function employed `exit()`, which, when invoked, results in the immediate termination of the entire program. In a library this is not a wanted response when handling errors, a better approach is to return an error code and let  the calling application handle the error.

(cherry picked from commit 477c9cd4f947936655feb3a5a3e38011052c1c25)
Signed-off-by: Ted Ralphs <ted@lehigh.edu>